Removing redundant code and Adding run class
I removed the main method from the BruteForceApproach class and created an apposite class to run the main method
In this class I let the user decide whether to input cities or take random cities, this is to facilitate testing and reviewing. For the same reason, I added some code to output execution and calculation time.